+++ src/quic_crypto_nif.erl
@@ -15,10 +15,19 @@
     new_hp_ctx/2,
     seal/4,
     open/5,
-    hp_block/2
+    hp_block/2,
+    protect_run/7
 ]).
 
--nifs([is_loaded/0, new_aead_ctx/3, new_hp_ctx/2, seal/4, open/5, hp_block/2]).
+-nifs([
+    is_loaded/0,
+    new_aead_ctx/3,
+    new_hp_ctx/2,
+    seal/4,
+    open/5,
+    hp_block/2,
+    protect_run/7
+]).
 -on_load(load/0).
 
 load() ->
@@ -68,3 +77,5 @@ seal(_Ctx, _Nonce, _AAD, _Plain) -> erlang:nif_error(not_loaded).
 open(_Ctx, _Nonce, _AAD, _CipherText, _Tag) -> erlang:nif_error(not_loaded).
 -spec hp_block(reference(), binary()) -> binary() | {error, atom()}.
 hp_block(_Ctx, _Sample) -> erlang:nif_error(not_loaded).
+protect_run(_AeadCtx, _HpCtx, _IV, _PN0, _FirstByteBase, _DCID, _Payloads) ->
+    {error, not_loaded}.